+++ components/layout/landing/hero-card.tsx
